backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
asyncpg==0.31.0
openai==2.21.0
python-dotenv==1.2.1
websockets==16.0
pgvector==0.4.2
